latest changes except the ones for development
