+++ DebsocWebsite/urls.py
@@ -25,5 +25,5 @@ from django.conf.urls import url
 urlpatterns = [
     path('', include('sitewebapp.urls')),
     path('admin/', admin.site.urls),
-    path('betaaudi/',include('audition.urls')),
+    path('audi/',include('audition.urls')),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
