+++ sitewebapp/templates/sitewebapp/base.html
@@ -136,12 +136,6 @@
             <li class="nav-item">
               <a class="nav-link" href="{% url 'about' %}">About</a>
             </li>
-            <li class="nav-item">
-              <a class="nav-link" href="{% url 'apd2' %}">APD II</a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link" href="{% url 'audition' %}">Audition</a>
-            </li>
           </ul>
         </div>
       </div>
