--- sitewebapp/templates/sitewebapp/showdata.html
@@ -1,8 +1,0 @@
-{% extends "sitewebapp/base.html" %} {% block content %}
-    {% for ans in answers %}
-        <h3>{{ans.q.roundno}} + {{ans.q.question}}</h3>
-        <p>{{ans.ans}}</p>
-        <h4>{{ans.anstime}}</h4/>
-        <br/>
-    {% endfor %}
-{% endblock %}
