+++ templates/allauth/account/email/email_confirmation_signup_subject.txt
@@ -1,0 +1,1 @@
+{% include "account/email/email_confirmation_subject.txt" %}
