+++ templates/allauth/account/messages/cannot_delete_primary_email.txt
@@ -1,0 +1,2 @@
+{% load i18n %}
+{% blocktrans %}You cannot remove your primary e-mail address ({{email}}).{% endblocktrans %}
