+++ templates/allauth/account/messages/email_confirmation_sent.txt
@@ -1,0 +1,2 @@
+{% load i18n %}
+{% blocktrans %}Confirmation e-mail sent to {{email}}.{% endblocktrans %}
