+++ templates/allauth/account/messages/email_confirmed.txt
@@ -1,0 +1,2 @@
+{% load i18n %}
+{% blocktrans %}You have confirmed {{email}}.{% endblocktrans %}
