+++ templates/allauth/account/messages/email_deleted.txt
@@ -1,0 +1,2 @@
+{% load i18n %}
+{% blocktrans %}Removed e-mail address {{email}}.{% endblocktrans %}
