+++ templates/allauth/account/messages/logged_out.txt
@@ -1,0 +1,2 @@
+{% load i18n %}
+{% blocktrans %}You have signed out.{% endblocktrans %}
