+++ templates/allauth/account/messages/password_changed.txt
@@ -1,0 +1,2 @@
+{% load i18n %}
+{% blocktrans %}Password successfully changed.{% endblocktrans %}
