+++ templates/allauth/account/messages/password_set.txt
@@ -1,0 +1,2 @@
+{% load i18n %}
+{% blocktrans %}Password successfully set.{% endblocktrans %}
