+++ templates/allauth/account/messages/primary_email_set.txt
@@ -1,0 +1,2 @@
+{% load i18n %}
+{% blocktrans %}Primary e-mail address set.{% endblocktrans %}
