+++ templates/allauth/account/messages/unverified_primary_email.txt
@@ -1,0 +1,2 @@
+{% load i18n %}
+{% blocktrans %}Your primary e-mail address must be verified.{% endblocktrans %}
