+++ templates/allauth/account/password_reset_from_key_done.html
@@ -1,0 +1,9 @@
+{% extends "account/base.html" %}
+
+{% load i18n %}
+{% block head_title %}{% trans "Change Password" %}{% endblock %}
+
+{% block content %}
+    <h1>{% trans "Change Password" %}</h1>
+    <p>{% trans 'Your password is now changed.' %}</p>
+{% endblock %}
