+++ templates/allauth/account/password_set.html
@@ -1,0 +1,15 @@
+{% extends "account/base.html" %}
+
+{% load i18n %}
+
+{% block head_title %}{% trans "Set Password" %}{% endblock %}
+
+{% block content %}
+    <h1>{% trans "Set Password" %}</h1>
+
+    <form method="POST" action="{% url 'account_set_password' %}" class="password_set">
+        {% csrf_token %}
+        {{ form.as_p }}
+        <input type="submit" name="action" value="{% trans 'Set Password' %}"/>
+    </form>
+{% endblock %}
